Move utils to same file
