--- internal/controller/utils/maps.go
@@ -1,38 +1,0 @@
-/*
-Copyright 2023.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package utils
-
-import "sort"
-
-type Mapping interface {
-	keys() []string
-}
-
-func GetValues(vm map[string]string) []string {
-	var keys []string
-	for k := range vm {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	return keys
-}
-
-func HasSameKeys(vm1, vm2 map[string]string) bool {
-	k1 := GetValues(vm1)
-	k2 := GetValues(vm2)
-	return SameList(k1, k2)
-}
